old-setup/backend/pyproject.toml
[tool.poetry]
name = "smartquery"
version = "0.1.0"
description = ""
authors = ["elilourens <elilourens123@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13.8"
langgraph = ">=0.5.3,<0.6.0"
langchain = ">=0.3.26,<0.4.0"
ollama = ">=0.5.1,<0.6.0"
pillow = ">=11.3.0,<12.0.0"
matplotlib = ">=3.10.3,<4.0.0"
open-clip-torch = ">=3.0.0,<4.0.0"
fastapi = ">=0.116.1,<0.117.0"
supabase = ">=2.17.0,<3.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
python-docx = ">=1.2.0,<2.0.0"
pymupdf = ">=1.26.3,<2.0.0"
transformers = "4.44.2"
uvicorn = {extras = ["standard"], version = ">=0.35.0,<0.36.0"}
sentence-transformers = "2.7.0"
python-jose = {extras = ["cryptography"], version = ">=3.5.0,<4.0.0"}
pinecone = ">=7.3.0,<8.0.0"
python-multipart = ">=0.0.20,<0.0.21"
langchain-community = ">=0.3.27,<0.4.0"
pyjwt = {extras = ["crypto"], version = ">=2.10.1,<3.0.0"}
cryptography = ">=45.0.6,<46.0.0"
nltk = ">=3.9.1,<4.0.0"
docx2txt = ">=0.9,<0.10"
langchain-openai = ">=0.3.33,<0.4.0"
numpy = ">=2.3.3,<3.0.0"
torch = "2.8.0"
torchvision = "0.23.0"
stripe = ">=11.4.0,<12.0.0"
celery = {extras = ["redis"], version = ">=5.3.0,<6.0.0"}
redis = ">=5.0.0,<6.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
